Fixed filter data.
