+++ red-river-mart/apps/frontend/src/apis/filterRepo.ts
@@ -1,0 +1,10 @@
+import type { FilterOption } from "../../../../shared/types/types";
+
+export const filterData: FilterOption[] = [
+  { "id": 1, "name": "Price: Low to High", "sort": "priceLowToHigh" },
+  { "id": 2, "name": "Price: High to Low", "sort": "priceHighToLow" },
+  { "id": 3, "name": "Electronics", "category": "electronics" },
+  { "id": 4, "name": "Clothing", "category": "clothing" },
+  { "id": 5, "name": "Books", "category": "books" },
+  { "id": 6, "name": "Accessories", "category": "accessories" }
+]
