+++ red-river-mart/apps/frontend/src/apis/forumRepo.ts
@@ -1,74 +1,68 @@
 import type { Forum } from "../../../../shared/types/types";
 
-export const forumData: Forum[] = [
-  {
-    id: 1,
-    subject: "co-housing",
-    title: "Looking for a Roommate",
-    date: "2023-11-26",
-    description: "I have an extra space for a roommate!",
-  },
-  {
-    id: 2,
-    subject: "jobs",
-    title: "Part-Time Developer",
-    date: "2024-01-15",
-    description: "Looking for a part-time React developer to join a small team.",
-  },
-  {
-    id: 3,
-    subject: "events",
-    title: "Community Cleanup",
-    date: "2024-02-10",
-    description: "Join us for a local park cleanup this weekend!",
-  },
-  {
-    id: 4,
-    subject: "housing",
-    title: "Apartment Available",
-    date: "2024-03-05",
-    description: "1-bedroom apartment available for rent in River Heights.",
-  },
-  {
-    id: 5,
-    subject: "buy/sell",
-    title: "Selling Used Laptop",
-    date: "2024-04-12",
-    description: "Lightly used MacBook Air for sale, excellent condition.",
-  },
-  {
-    id: 6,
-    subject: "community",
-    title: "Volunteers Needed",
-    date: "2024-05-20",
-    description: "Looking for volunteers to help with a food drive event.",
-  },
-  {
-    id: 7,
-    subject: "jobs",
-    title: "Freelance Designer Wanted",
-    date: "2024-06-14",
-    description: "Seeking a graphic designer for a short-term branding project.",
-  },
-  {
-    id: 8,
-    subject: "housing",
-    title: "Sublet Available",
-    date: "2024-07-22",
-    description: "Fully furnished sublet available from July to September.",
-  },
-  {
-    id: 9,
-    subject: "events",
-    title: "Coding Workshop",
-    date: "2024-08-30",
-    description: "Free beginner coding workshop hosted at the community center.",
-  },
-  {
-    id: 10,
-    subject: "buy/sell",
-    title: "Furniture Giveaway",
-    date: "2024-09-10",
-    description: "Giving away a couch and coffee table, must pick up.",
+type ForumsResponseJSON = { message: string; data: Forum[] };
+type ForumResponseJSON = { message: string; data: Forum };
+
+const BASE_URL = `${import.meta.env.VITE_API_BASE_URL}/api/v1/forums`;
+
+export async function getAllForums(): Promise<Forum[]> {
+  const response: Response = await fetch(BASE_URL);
+
+  if (!response.ok) {
+    throw new Error("Failed to fetch the forums");
   }
-];
+
+  const json: ForumsResponseJSON = await response.json();
+  return json.data;
+}
+
+export async function getForumById(id: number): Promise<Forum> {
+  const response: Response = await fetch(`${BASE_URL}/${id}`);
+
+  if (!response.ok) {
+    throw new Error(`Failed to fetch the forum with id ${id}`);
+  }
+
+  const json: ForumResponseJSON = await response.json();
+  return json.data;
+}
+
+export async function createForum(forum: Omit<Forum, "id">): Promise<Forum> {
+  const response: Response = await fetch(BASE_URL, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(forum),
+  });
+
+  if (!response.ok) {
+    throw new Error("Failed to create the forum");
+  }
+
+  const json: ForumResponseJSON = await response.json();
+  return json.data;
+}
+
+export async function updateForum(forum: Forum): Promise<Forum> {
+  const response: Response = await fetch(`${BASE_URL}/${forum.id}`, {
+    method: "PUT",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify(forum),
+  });
+
+  if (!response.ok) {
+    throw new Error(`Failed to update forum with the id ${forum.id}`);
+  }
+
+  const json: ForumResponseJSON = await response.json();
+  return json.data;
+}
+
+export async function deleteForum(id: number): Promise<void> {
+  const response: Response = await fetch(`${BASE_URL}/${id}`, {
+    method: "DELETE",
+  });
+
+  if (!response.ok) {
+    throw new Error(`Failed to delete forum with the id ${id}`);
+  }
+}
