--- red-river-mart/apps/frontend/src/apis/mockItemRepo.ts
@@ -1,245 +1,0 @@
-import type { Item } from "../../../../shared/types/types";
-import placeholderImage from "../assets/placeholder.png";
-
-export const itemData: Item[] = [
-  {
-    id: 1,
-    name: "USB Flash Drive 32GB",
-    category: "electronics",
-    price: 25.0,
-    src: placeholderImage,
-    description: "Brand new USB 32GB flash drive, still in packaging. $25 OBO."
-  },
-  {
-    id: 2,
-    name: "Wireless Mouse",
-    category: "electronics",
-    price: 18.5,
-    src: placeholderImage,
-    description: "Lightly used wireless mouse, works perfectly. $18 OBO."
-  },
-  {
-    id: 3,
-    name: "Portable Power Bank",
-    category: "electronics",
-    price: 30.0,
-    src: placeholderImage,
-    description: "Portable power bank, barely used, charging cables included. $30 OBO."
-  },
-  {
-    id: 4,
-    name: "Headphones",
-    category: "electronics",
-    price: 45.0,
-    src: placeholderImage,
-    description: "Over-ear headphones, lightly used, great sound. $45 OBO."
-  },
-  {
-    id: 5,
-    name: "School Hoodie",
-    category: "clothing",
-    price: 35.0,
-    src: placeholderImage,
-    description: "Brand new hoodie, never worn. Perfect for chilly days. $35 OBO."
-  },
-  {
-    id: 6,
-    name: "Graphic T-Shirt",
-    category: "clothing",
-    price: 20.0,
-    src: placeholderImage,
-    description: "Lightly worn graphic tee, still looks new. $20 OBO."
-  },
-  {
-    id: 7,
-    name: "Baseball Cap",
-    category: "clothing",
-    price: 15.0,
-    src: placeholderImage,
-    description: "Classic cap, lightly used, great condition. $15 OBO."
-  },
-  {
-    id: 8,
-    name: "Winter Jacket",
-    category: "clothing",
-    price: 60.0,
-    src: placeholderImage,
-    description: "Warm winter jacket, brand new with tags. $60 OBO."
-  },
-  {
-    id: 9,
-    name: "Math Textbook",
-    category: "books",
-    price: 55.0,
-    src: placeholderImage,
-    description: "Math 101 textbook, lightly highlighted, like new. $55 OBO."
-  },
-  {
-    id: 10,
-    name: "English Literature Book",
-    category: "books",
-    price: 40.0,
-    src: placeholderImage,
-    description: "Lit textbook, some notes inside but in great shape. $40 OBO."
-  },
-  {
-    id: 11,
-    name: "Journal/Planner",
-    category: "books",
-    price: 15.0,
-    src: placeholderImage,
-    description: "Planner barely used, most pages blank. $15 OBO."
-  },
-  {
-    id: 12,
-    name: "Sketchbook",
-    category: "books",
-    price: 12.0,
-    src: placeholderImage,
-    description: "Sketchbook, some doodles inside, mostly blank. $12 OBO."
-  },
-  {
-    id: 13,
-    name: "Reusable Water Bottle",
-    category: "accessories",
-    price: 10.0,
-    src: placeholderImage,
-    description: "Brand new water bottle, eco-friendly and leak-proof. $10 OBO."
-  },
-  {
-    id: 14,
-    name: "Keychain Lanyard",
-    category: "accessories",
-    price: 5.0,
-    src: placeholderImage,
-    description: "Cute lanyard for keys or ID, lightly used. $5 OBO."
-  },
-  {
-    id: 15,
-    name: "Backpack",
-    category: "accessories",
-    price: 45.0,
-    src: placeholderImage,
-    description: "Sturdy backpack, lightly used, perfect for school or work. $45 OBO."
-  },
-  {
-    id: 16,
-    name: "Phone Case",
-    category: "accessories",
-    price: 12.5,
-    src: placeholderImage,
-    description: "Brand new phone case, never used. $12 OBO."
-  },
-  {
-    id: 17,
-    name: "Laptop Sleeve",
-    category: "electronics",
-    price: 25.0,
-    src: placeholderImage,
-    description: "Lightly used laptop sleeve, protects your laptop perfectly. $25 OBO."
-  },
-  {
-    id: 18,
-    name: "Smartwatch",
-    category: "electronics",
-    price: 150.0,
-    src: placeholderImage,
-    description: "Smartwatch, barely used, works like new. $150 OBO."
-  },
-  {
-    id: 19,
-    name: "USB-C Cable",
-    category: "electronics",
-    price: 8.0,
-    src: placeholderImage,
-    description: "Extra USB-C cable, brand new in packaging. $8 OBO."
-  },
-  {
-    id: 20,
-    name: "Denim Jeans",
-    category: "clothing",
-    price: 45.0,
-    src: placeholderImage,
-    description: "Denim jeans, lightly worn, still look new. $45 OBO."
-  },
-  {
-    id: 21,
-    name: "Winter Scarf",
-    category: "clothing",
-    price: 22.0,
-    src: placeholderImage,
-    description: "Cozy scarf, lightly used, perfect for winter. $22 OBO."
-  },
-  {
-    id: 22,
-    name: "Gym Shorts",
-    category: "clothing",
-    price: 28.0,
-    src: placeholderImage,
-    description: "Comfortable gym shorts, barely worn. $28 OBO."
-  },
-  {
-    id: 23,
-    name: "Fleece Joggers",
-    category: "clothing",
-    price: 38.0,
-    src: placeholderImage,
-    description: "Soft fleece joggers, lightly used. $38 OBO."
-  },
-  {
-    id: 24,
-    name: "Science Textbook",
-    category: "books",
-    price: 65.0,
-    src: placeholderImage,
-    description: "Science textbook, lightly highlighted, excellent condition. $65 OBO."
-  },
-  {
-    id: 25,
-    name: "Fantasy Novel",
-    category: "books",
-    price: 18.0,
-    src: placeholderImage,
-    description: "Fun fantasy novel, brand new. $18 OBO."
-  },
-  {
-    id: 26,
-    name: "Cookbook",
-    category: "books",
-    price: 25.0,
-    src: placeholderImage,
-    description: "Cookbook lightly used, some notes inside. $25 OBO."
-  },
-  {
-    id: 27,
-    name: "Digital Stylus",
-    category: "accessories",
-    price: 50.0,
-    src: placeholderImage,
-    description: "Digital stylus for notes or art, barely used. $50 OBO."
-  },
-  {
-    id: 28,
-    name: "Laptop Stand",
-    category: "accessories",
-    price: 35.0,
-    src: placeholderImage,
-    description: "Ergonomic laptop stand, lightly used. $35 OBO."
-  },
-  {
-    id: 29,
-    name: "Sunglasses",
-    category: "accessories",
-    price: 20.0,
-    src: placeholderImage,
-    description: "Sunglasses, brand new, perfect for sunny days. $20 OBO."
-  },
-  {
-    id: 30,
-    name: "Notebook",
-    category: "books",
-    price: 8.0,
-    src: placeholderImage,
-    description: "Notebook, mostly blank pages, lightly used. $8 OBO."
-  }
-];
