--- red-river-mart/apps/frontend/src/apis/userMockRepo.ts
@@ -1,84 +1,0 @@
-import type { User } from "../../../../shared/types/types";
-
-export const userData: User[] = [
-    {
-    "id": 1,
-    "userName": "Ranindi Gunasekera",
-    "bio": "Hello! I am a student in the AD&D program and I am interested in buying and selling computer parts",
-    "email": "ranindi@rrc.ca",
-    "phone": "657-576-3756",
-    "preferredContact": "Email/ Teams"
-    },
-    {
-    "id": 2,
-    "userName": "Heuone Castillo",
-    "bio": "Tech enthusiast and gamer. Always looking for deals on PC parts and accessories.",
-    "email": "heuone@rrc.ca",
-    "phone": "645-433-6785",
-    "preferredContact": "Teams"
-    },
-    {
-    "id": 3,
-    "userName": "Alex Richard Carlos",
-    "bio": "Jewelry maker and small business owner sharing handmade accessories and gifts.",
-    "email": "alex3@rrc.ca",
-    "phone": "364-563-4875",
-    "preferredContact": "Email"
-    },
-    {
-    "id": 4,
-    "userName": "Jackson Odudu-Abasi",
-    "bio": "Book lover curating a collection of novels and art prints for sale.",
-    "email": "odudu-abasi@rrc.ca",
-    "phone": "233-434-3423",
-    "preferredContact": "Email/ In-app"
-    },
-    {
-    "id": 5,
-    "userName": "Renz Tengco",
-    "bio": "Casual seller focused on sneakers, sportswear, and fitness equipment.",
-    "email": "renzt@rrc.ca",
-    "phone": "345-434-5432",
-    "preferredContact": "Phone"
-    },
-    {
-    "id": 6,
-    "userName": "Carlos Boquia",
-    "bio": "Outdoor enthusiast who frequently sells and trades camping and hiking gear.",
-    "email": "carlosbq@rrc.ca",
-    "phone": "657-343-3756",
-    "preferredContact": "Teams/Phone"
-    },
-    {
-    "id": 7,
-    "userName": "Zizhang He",
-    "bio": "Bookworm clearing out shelves to make space for new reads — all genres welcome!",
-    "email": "zizhang@rrc.ca",
-    "phone": "345-435-4342",
-    "preferredContact": "Email/ Teams/ Phone"
-    },
-    {
-    "id": 8,
-    "userName": "Dwayne Malit",
-    "bio": "DIY enthusiast selling tools, materials, and unfinished projects to fellow makers.",
-    "email": "dmalit@rrc.ca",
-    "phone": "343-576-3756",
-    "preferredContact": "Email"
-    },
-    {
-    "id": 9,
-    "userName": "Kavya Kiran",
-    "bio": "Vintage collector with a passion for home décor and timeless clothing pieces.",
-    "email": "kavyak@rrc.ca",
-    "phone": "657-342-3485",
-    "preferredContact": "Teams/ Phone"
-    },
-    {
-    "id": 10,
-    "userName": "Babou Jobe",
-    "bio": "Student artist selling prints, sketchbooks, and art supplies.",
-    "email": "bjobe@rrc.ca",
-    "phone": "343-455-3756",
-    "preferredContact": "Email/ Teams"
-    }
-]
