--- red-river-mart/apps/frontend/src/service/userService.ts
@@ -1,23 +1,0 @@
-// import * as UserRepo from "../apis/userRepo";
-// import { User } from "@shared/types/types";
-
-// export async function fetchUsers(): Promise<User[]> {
-//     return UserRepo.fetchUsers();
-// }
-
-// export async function getUserById(userId: string): Promise<User> {
-//     return UserRepo.getUserById(userId);
-// }
-
-// export async function updateUser(user: User, sessionToken?: string|null): Promise<User> {
-//     return UserRepo.updateUser(user, sessionToken);
-// }
-
-// export async function createUser(user: Omit<User, "id">): Promise<User> {
-//     return UserRepo.createUser(user);
-// }
-
-// export async function deleteUser(userId: string): Promise<void> {
-//     return UserRepo.deleteUser(userId)
-// }
-
